requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.3.1
diffusers==0.27.2
transformers==4.36.2
opencv-python==4.10.0
safetensors==0.4.3
pillow==10.2.0
einops==0.8.0
peft==0.11.1
protobuf==3.20
streamlit==1.36.0